Replace PlayerVsPlayer component and add back icon
The PlayerVsPlayer component for the "/pvp/online" path in App.tsx has been replaced with a SignIn component, shifting the usage of PlayerVsPlayer to a more specific context. Additionally, a 'back' SVG icon has been added to the project's assets and implemented in the PVPMenu component. The routes in the PVPMenu have also been adjusted accordingly.