Add TODO comment for default gameboard state in online mode
A TODO comment has been added in the GameBoard component to set a default empty state for the gameboard, specifically for the online game mode. This will ensure that the gameboard is correctly initialized and uploaded to Firebase before rendering, thus preventing any discrepancies when starting a new game online.